fix(app): retrieve initial route name at mount
The React app instance might not be initialized in time to receive an event from the Android fragment. Compensate this issue by implementing an imperative asynchronous data retrieval when the component mounts for the first time. This allows the app to still be able to get the initial route, albeit at a noticeable yet acceptable UI delay.